+++ changelog.md
@@ -16,10 +16,12 @@ Codion Change Log
 - TableSelectionModel.beforeSelectionChangeEvent() renamed selectionChangingEvent().
 ### is.codion.swing.common.ui
 - FilteredTable.copyRowsAsTabDelimitedString() replaced with copyToClipboard() and copySelectedToClipboard().
+- Controls.empty() now ignores separators.
 ### is.codion.swing.framework.ui
 - EntityTablePanel.TableControls.COPY_TABLE_DATA renamed COPY, COPY_CELL and COPY_ROWS added.
 - EntityTablePanel.TableControls.COPY, EDIT_ATTRIBUTE, CONFIGURE_COLUMNS renamed COPY_CONTROLS, EDIT_ATTRIBUTE_CONTROLS and COLUMN_CONTROLS respectively, PRINT_CONTROLS, CONDITION_CONTROLS and FILTER_CONTROLS added.
 - EntityTablePanel controls configuration refactored.
+- EntityTablePanel.configurePopupMenu() and configureToolBar() added.
 
 ## 0.17.42
 ### is.codion.common.core